Implement 'reports' tab for a grant
This adds a 'Reports' tab to the service navigation when viewing a specific
grant. This is the home for monitoring reports for the grant. It serves two sets
of users: grant team members (who will be able to view the reports and test
them), and grant admins (who will be able to create/manage/delete reports in the
future).

Designs for the end-to-end reports tab are still up in the air, so we're
building a subset of the journey, mostly focused around allowing grant team
members to view and test the reports. However we'll let grant admins create and
manage reports and tasks (importantly not questions yet, as those are the
designs/part of the journey still being firmed up).

This patch lists all collections for the grant (currently we are only creating
monitorig reports). All of the links on the page are dead links and will be made
functional in future work.