Make alembic-postgresql-enum a core dependency
We use this dependency to generate DB migrations for changes to enum types. I
thought we'd only need this as a dev dependency to generate the migration, but
we've just had an error on deployment after removing dev dependencies from the
deployed package.

This is because we import it in `env.py`, which is loaded when running database
migrations.