FSPT-721: manage collections
Adds a page to the Reports tab for 'managing' a collection: specifically
renaming and deleting it. We aren't super confident in the terminology here, but
will go with it and adjust based on feedback if we need to.

We haven't yet dealt with collection versioning, which we need to, but we try to
anticipate some production-based problems by making it impossible to delete a
collection (version) that has associated live submissions.

It's possible that we should actually block all changes to a collection that's
gone live, but we don't have the concept of collection versioning built in yet.
So there's bits to revisit here.