FSPT-723: create 'add task' page
Create, and hook up (for a report with 0 tasks), the 'add task' page. This takes
the name for a task and creates the task in the report.

For now, it returns the user back to the 'list reports' page. The following
patch will add a 'list tasks in a report' dashboard-style page, and 'add task'
will then redirect there instead.