FSPT-727: Add fallback validator for checkbox final options
We use the GOV.UK checkbox component 'exclusive' for final option choice
to deselect any other choices and vice versa when the user completes
the form.

However this relies on JS being enabled, so we now add a custom validator
to still ensure a user can't submit when selecting both the separated
final checkbox option and another checkbox.