FSPT-724: manage tasks in a report
Adds and links up a page where you can "manage" a form - rename and delete it.
This lines up with the "manage report" page that lets you do the same functions.

This page is available to grant admins and above only, and only when there are
no live submissions against the report yet. We want to consider report schemas
immutable once submissions have started to come in against them, and plan future
work to adding versioning that will let changes be made if need be.