FSPT-727: Add Specifically mgd expression for checkboxes
Adds support for a new managed expression type, 'Specifically', for
use with Checkbox questions. Useable for conditional routing to allow
the form designer to select exactly one option from a list the list of
data source items from a checkbox question type - if the user selects
that option (singly or amongst other selected options) then the
expression evaluates to True and the conditional question is shown.