Tweak page designs
After a review with Brian this morning, we've settled on the following changes
and have an idea of how they'll flow through to the question pages, so we can
start to build out again.

Change name and delete a report have been lifted back to top-level card actions.

Change task name and delete a task have been nested inside the 'manage task'
page, where you do all of: renaming task, adding+managing questions, and
deleting the task. This allows the task summary card to not have a load of dense
actions.