Bit of thinking about surfacing grant names
We currently show the grant name as a heading caption on a bunch of pages, but
on pages that need a different caption we're struggling to get the grant name
showing up in a consistent and nice place. This does a bit of thinking around
how we could show the grant name + navigation in a single consistent place,
removing some heading captions and freeing it up for other uses.