FSPT-727: Add and update tests for report previews
As well as updating existing tests to account for previewing reports and
tasks from the Reports tab, now we also:
- extensively test the new report preview routes
- add permissions check to the form runner to only allow users to preview
their own submission. This will need to be extended when we work on the
Access Grant Funding live submissions to only users from the same
Organisation to be able to edit a live submission, rather than a per-user
basis.