Fix submission status on 'list submissions' page
This was showing the 'status' field of the Submission DB model, rather than the
generated status from the submission helper.

We also remove the 'status' column from the Submission table, since this isn't
actually used everywhere. For now, we dynamically determine the status based
on submission events and form data.