FSPT-605: allow adding flat questions to groups
Starts hooking into the parent of adding components by allowing an "Add
question" link on a group to include the parent relationship.

Note with this commit we're limited to only adding children to one level
of group nesting but with the manage group pages we'll be able to add
children to children of groups.