FSPT-744: add+edit conditions and validation
Lifts these pages across from the developers tab and implements them in the
Reports tab, which will be available for all (grant) users. These pages
specifically will only be available to grant admins initially, which will be
form designers in the Funding Service.

We expect to tweak the designs of these pages over the next week or so in
collaboration with the design team. These pages have been through user testing
and held up reasonably well, but we expect to make some tweaks and improvements
regardless.