FSPT-605: remove nested group check and cover conditions
This was tried out while spiking non-realistic scenarios and thought
it was working appropriately but we'll remove this and optimise later if
needed.

The intention was to only selectin for components that are part of
groups and avoid the selectin for all of the top level form components
(which you might assume are the majority).

Remove this for now and we'll see if there's any need to optimise
as the first realistic forms get modelled.