FSPT-606: add end to end steps for adding a group
This uses the existing questions to test methodology but allows those
questions to be included in groups to cover both the form runner routing
to group questions and the form builder being able to add groups.

When we need to check more complex things like group conditions we may
need to revise this interface but this should do for now (those things
are likely covered lower down anyway).