Save and render prefixes and suffixes for numbers
We recently added support for prefixes and suffixes in form building for integer
fields, so you can for example prepend the input with a '£' for currency inputs.

I forgot to cascade this through to places where those values are rendered out,
for example the check your answers page and the CSV export. This patch addresses
  that missing functionality.

It also formats numbers with commas to separate thousands/millions/etc.

If we had real data we'd have to consider how to manage doing migrations for
existing submissions, but right now we can still just empty everything out -
fortunately.