Prevent prefix+suffix from storing empty strings
We exclude prefix and suffix from serialising out to the DB if the values are
None, but we'll still store them if the values are empty strings.

(HTML) form submissions send these fields across as empty strings, so we've been
serialising them. Let's add some filters to coerce these to proper Nones so that
we can drop them out of the DB json blob if they're effectively unset.