FSPT-604: manage group display options page
Adds a page to allow form designers to configure the presentation
options on a group. This uses the same interface as the existing
presentation options for questions but for the time being only has one
option.

If we feel like its worth splitting it out at this stage that would also
be possible but just adding this on does keep the typing simple.

This will then be exposed by a property that allows question form
building logic to show all the components in this group at once or
continue to route normally for one question per page.