FSPT-751: list group questions page
Adds the detail/ manage page to click in to manage question groups. For
now this supports listing the questions in a group and deleting the
group, it updates the page join logic to factor in groups:
- adding a question will return you to the task or group you're in
- moving a question up and down will keep you were you performed that
  action
- breadcrumbs allow you to navigate from the group up to the task

This also follows the same delete component pattern as deleting
questions and tasks, a future commit should make sure any integrity
checks that are performed to let you delete are updated to factor in
checking questions within groups.