FSPT-751: add group pages
This commit adds the relevant pages needed for the add question group
flow. It was locked in during a design/ dev mobbing session that in
order to add a group you would both choose its name and if it should be
shown on the same page or not.

To keep the change set clean a future commit will add the presentation
options for groups to actually be set.

A future commit will allow the name and display options that are set
during creation to be changed.