FSPT-747: Replicate dev e2e test for Reports tab
We have an existing create & preview report test covering the Developers
tab journey. This creates the same baseline journey for the Reports
tab along with all the page helpers and locators for the Reports tab
views.

This drops the testing of question groups as that is not yet present
in the reports tab, but we'll add them back in later once merged.