FSPT-751: add groups flow blocks duplicate names
As the add question group flow first asks for the group name and then
for the groups presentation option it currently doesn't nicely in-line
validate if the name hasn't already been taken for that form. Check the
name as its being provided to avoid the user getting further in the flow
and running into an unmanaged error.