FSPT-747: Fail e2e test if new question type added
Adds an assertion to fail the create & preview collection e2e test if a
new question type is added but not covered in the e2e test.

We raise the error after the test clean-up so that, if the assertions
fail, the test grant/report etc. is still deleted rather than being
blocked halfway through the test from the failed assertion.