FSPT-751: update integrity checks to compare moving groups
Now that the components moving up and down and we groups of components
we need to check two things:
- the group we're moving itself doesn't have a condition on whats being
  swapped
- whats being swapped doesn't have a condition on _anything_ in the
  group

This extends the existing integrity checks and instead of checking the
component and swap components conditions directly it checks for any
component that might be inside whats being moved.

This would likely benefit from more thorough coverage but this starts
with scenarios that should help avoiding regression.