FSPT-751: update component order checks to use the global schema order
We previously relied on checking the `order` property of components to
check if you could depend on a question that came above the one having a
condition applied to. Update this to use the form components order after
having flattened all nested component relationships.

Fixed up end to end tests finding multiple links with group names.