BAU: Remove pending invitations for platform admins
During user testing we added a user as a grant team member but
subsequently had to add them to the FSD_ADMIN group and get them to sign
in as Funding Service Admins. This left the pending Grant Team member
invitation unclaimed and made the Grant Team view look odd (with them
appearing in both tables).

This is a small edge case but when a funding service admin logs in, as
well as removing any existing roles they might have, we should also
claim any pending invitations.