FSPT-715: only manage guidance text on question groups when same page
This commit limits the ability to manage guidance on the question group
if its same page, with the edit question page steering you towards the
group.

For now we don't limit behaviour otherwise which allows you to move back
and forwards from same page/ guidance on question pages (without
removing it or requiring the user to remove it).

We can review the behaviour here and see if we want to add any warnings
or quality of life improvements.