FSPT-753: link forms to collections directly
We no longer use the concept of sections to group related forms together. We
initially thought this might be useful but our current focus of monitoring +
evaluation reporting doesn't use them. 

Let's remove them from the codebase to keep our current domain understanding
aligned with the codebase.

This is the first of many PRs to incrementally remove the concept. We start by
adding the link directly from the Form model to the Collection model, but all of
the code still understands and relies on the Form -> Section -> Collection flow.
Subsequent PRs will start to unpick that.