BAU: cache commonly accessed form properties for helpers
We're assuming that form/ question changes happen at the end of HTTP
requests and that its OK to have the same calculated property returned
throughout the lifecycle of one request.

This is an assumption we'll need to validate and make clear - we've
started by prefixing any properties that are cached with `cached_` to
make it clear to devs that if they change something they'll probably
still get the old result.

Update the tests to clear cache after accessing those properties.

We've observed just these changes move us from the tasklist loading at
~650ms to ~280ms which indicates a lot of time was being spent
recursively calculated copmonents in the CPU.