BAU: add lru caches to method
We only want the cache to apply on the instances of the helpers created
which the lru_cache decorator will not do by default.

If we apply the decorator to the methods on these classes the results
will be cached along with the "self" instance of that class as one of
the called properties - this means the results will never be removed
from the cache as that instance will never exist again and can quickly
lead to memory leak issues.

We could write a more generic wrapper to do instanced caching with
variable params but for now just provide a cached verison on the class.