FSPT-761: dont lazy load expressions on form pages
Asserts that getting a submission will not lazy load when expressions
are interrogated.

Snapshot:
- a small form tasklist will now load in ~20ms and a larger form in
  ~80ms.

It still doesn't feel like a comfortable gap as I think we should be
able to load much more complex and longer forms so I don't think this is
sufficient - we may need to optimise further looking into how the ORM is
used and how much of the schema we're loading at once.