FIX: preview task back link should return you to the form builder
As this is only ever needed by DGF previewing we've just added an
optional override to the "before" content back link - this means we can
continue to let the runner decide if the preview banner should be shown
etc.

Only the deliver grant funding will decide if an override should be
passed in based on if the session indicates we're previewing a task
specfically.