FSPT-760: fix sync to dev pipeline
We have a scheduled job that deploys the main branch to the dev
enivoronment to keep it up to date for developers to throw things at the
environment but keep ti separate from the core deployment path.

When we moved our primary image build to a shared ECR we also made the
image tags immutable which seems to cause paketo to fail on publish.
There may be build pack specific ways of instructing publishing that
this is the case (as it already seems to check the manifest for the
image at the start) but for now just ignore the package and build step
if its already been run.

We'd only expect to run package for the same git SHA if its a scheduled
job thats running sepearately and later for the dev environment and its
already been run for the deployment pipeline - this shouldn't be a
common occurence.

Example failure mode:

```
PUT
TAG_INVALID: The image tag '0f1d07c932342d2766a9ca9565244f80820d20d9'
already exists in the 'funding-service-ci-shared-app' repository and
cannot be overwritten because the tag is immutable.
```