FSPT-747: Add Question Groups to reports tab e2e tests
This ended up being a more significant rejig of the e2e test than
anticipated because we now we need to cater to question groups which
have their own guidance & display options and also nested questions.

We want the question creation to be able to continue to work for groups
or for flat questions, so this required a few additional checks and
providing some context as to where you're adding questions from, within
a group or within a task.

Question groups also change how we check visibility of elements on a
certain pages in the form runner, especially with guidance and when
all questions are displayed on one page.