Make notification banners blue again
When we set `govuk-brand-colour` to MHCLG teal, we did it with the intention of
changing only the header and footer bars. However, the current Design System
implementation for notification banners also uses the `govuk-brand-colour`, so
our notification banners now have teal borders as well. 

We don't like the look of this, and prefer them to stay blue for the standard
case. So here we add some overrides back on those classes to restore them to
GOV.UK Blue.

When the Design System roll out a fix for this issue, we can remove this
workaround.