Allowing managed expressions to specify functions that can be called in their statement (#748)
* Adding an abstract function called required_functions to the definition of ManagedExpression.
This allows a managed expression to use additional functions in the statement to be evaluated without falling foul of the locking down of simple eval. For example, if you need to use a custom function to determine something that is used in your statement, you can register this function as being allowed for that type of managed expression. See TestEvaluatingManagedExpressionsWithRequiredFunctions for examples. This will be used when we support a Date question type

* Adding default implementation for required_functions()

* Adding required_functions property to Expression model to simplify retrieving the required_functions for an expression. Updated tests accordingly