Interpolate context in question text/hints/guidance
As of this commit, there's no (nice) way for a form designer to actually put
context into the question text/hint/guidance, but if they do, then this patch
will properly resolve that context to the appropriate value in the form runner
and form designer parts of the service.

The format for interpolating text is to write the context reference inside
double parens (( )). For now, the context available is primarily answers to
questions in the submission, where the reference is the `safe_qid` of the
question. So a question text with interpolation might look like: 

What is the ((q_058d7d3a3e5e4b949a8157882961c59a))'s email address?