Single place to build an ExpressionContext
And rename expression_context->evaluation_context, since we'll soon have a
separate context for interpolation that includes some fallbacks to question
names. This only makes sense for interpolating text, not evaluating expressions
like a > b.

We want to be able to pull all of the relevant context for an expression
together in one place so that it can be used across various parts of the
codebase, including:

- on form runner pages (question, check your answer, etc) 

- on form designer pages (list task questions)

These have different amounts of context available; only form runner has a
submission associated and can inject actual user answers. Form designer needs to
be able to still interpolate text and inject context, but may fall back on
default text when no answers are available.