Micro-journey for adding context to a question
Adds a step-by-step journey to help a form designer insert a reference to some
dynamic context into a question's text or hint. In both of these fields there is
a new 'Insert data' button which initiates the journey. The user is first told
to select the 'bucket' of data they want to use, and then from that bucket they
choose the speific data point. For now the only 'bucket' is the set of questions
in the current task, and then they choose the specific question.

A future PR will constrain the list of questions to only 'earlier' questions,
for data integrity/logical consistency.

---

Move `insert data` button inside text/hint field

The button looked quite detached and unrelated when nestled below the inputs for
question text and question hints. This patch changes the `question name` input
to a textarea (any reasonably complex question was already being aggressively
truncated by the length of this box), and then moves the `Use data` button to
the right-hand lower corner of the input, which feels like it provides a
stronger tie to what the action is doing.