Make the default submit save, rather than add data
When pressing enter in a text input field, it will trigger form submission with
the first button in the form. This was one of the 'Add data' buttons, which
definitely isn't the intuitive action here.

We duplicate the 'add/save question' button at the top of the form so that it's
the default submit action. We fully hide it both visually and from screen
readers so as not to be presented as an actual button to anyone.