Prevent enter/newlines in question text+hint boxes
These will only ever display on a single line, and so we shouldn't let form
designers think they can enter multiple lines.

We use textareas instead of single-line text inputs simply because of the amount
of text they are likely to write in these boxes. The limited width of text
inputs means that text is often hidden out of bounds, which isn't a nice form
designing experience.