Move `markdown_to_html` into deliver_grant_funding
We use this endpoint to render a preview of guidance for a form designer. Now
that we can insert context into the guidance, we need to be able to render the
guidance with that context interpolated. This means we need to expose the
grant/collection context to the API endpoint, which means that we need to apply
further permission checks (such as: does the user making the request have access
to that grant).

This no longer feels like a general-purpose API endpoint, so moving it into
Deliver grant funding.