Only wipe context session data on save
Using the browser back button to navigate back into the 'add data' flow is
broken if we wipe the session on the first page load after ending that journey.

We retain the session information now until a save on the question (either
adding the question, editing it, or saving its guidance). 

To prevent applying data to the wrong question, we check if we're on an add
question page (with data from an add question page, where question_id=None in
the session) or an edit question page (where the question_id being edited must
match the question_id in the session).