Add table for component references
This table will be used when a component or an expression refers to another
component. By storing these references in the DB we can ensure that relied-upon
components cannot be deleted if there are references to them.

Some examples of this will be:

The question text for q2 needs to show the answer from q1; we will add a
ComponentReference from q2 to q1.

An expression on q2 uses an answer from q1 to determine whether q2 should be
shown; we'll add a ComponentReference from q2 (and its expression) to q1.

We'll record both the component and the expression (if appropriate) where the
reference originates from.

When updating a component (create/edit/delete) or expressions
(create/edit/delete), denormalise any references to another component so that we
can have the DB protect these references through FKs.