Check for valid expression references
When creating or updating questions and question groups, use a real expression
context to work out what variables are actually available for use.

For now this is primarily just the set of questions in the collection, but in
the future this might mean exposing things like the grant, grant recipient
profile, previous collections, or other sources of context.

This allows us to validate that any references used can actually be resolved to
a real piece of data, preventing form designers from entering invalid
references.