Fix tests and tighten up assertions
While adding some spies to add/edit_question to make sure that the validation
and sync of component references happens, I noticed that some of our tests have
been silently failing since adding the journey for inserting data. When
instantiating a form in a test and using `form.data`, submit fields are included
with their value set to boolean `False`, but this gets serialised to the string
`False` by client.post, which ends up then being treated as truthy when the
request handler's form processes it. So those endpoints were then going down the
'insert data' journey instead of saving the question.

Fix this up and then add a consistent way to extract data from a form for
passing to a test client's post method.

Also adds a couple of tests to the add+edit question pages for starting the
insert context journey.