Prevent breaking referential integrity between components
Now that components (questions and groups) can include answers from other
questions in their text, we need to make it impossible to invalidate that
reference after it's been set up. One of the ways this is possible is by setting
up a reference from q2 to q1, and then making q2 be shown to a user before q1.

This logic already exists in the function `_check_component_order_dependency`,
which is called when moving components up and down.

By changing the logic for this function to scan our new ComponentReference table
instead of checking just the conditions on each component, we can achieve this
check in a more reliable way. ComponentReferences are created whenever a
question refers to another one, either in its text or through its expressions
(conditions and validation).