Prevent using questions from later in the task
When inserting a reference to some contextual data, we currently let you add a
reference to a question that is displayed after the one being edited.

Referentially this doesn't make sense, so let's not list those questions in the
select dropdown.

Once the reference is set up, other integrity checks will prevent it from being
broken (eg by re-ordering questions)