FSPT-780: Prevent same-paging questions that refer to each other
If two questions refer to each other in their text/hint/guidance/etc, it's a
logical error for them to be able to display on the same page - we won't have an
answer for one of the questions to inject into the other.

This patch protects against creating references between questions in a same-page
group, and from changing a group to display on the same page if two questions in
it refer to each other.

We also prevent any questions in a same-page group from appearing in the 'insert
data' micro-journey question dropdown.