Change check-your-answers page to full width
The check your answers summary list looks overly cramped and dense when in a
two-thirds column. Given that it it limits the width of each column to less than
a standard two-thirds, we should be able to increase the width of this component
to full so that the question and answers have a little bit more room to spread
out, which should generally increase readability.