BAU: Don't allow questions to reference themselves
When inserting data from other questions into a question's text, hint
or guidance we don't surface the current question in the flow, but
a user could manually add a reference to itself we weren't catching it.

This change just raises an error if a user tries to manually reference
the current question, as this wouldn't make sense in the context of the
form and would look broken for grant recipients.