Remove code use of data source item references
This stops all of our code from using the DataSourceItemReference table in
preparation for deleting it altogether. All code now inspects the
ComponentReference table instead, looking at depends_on_data_source_item_id as
needed.

When we merge and deploy this, we'll be able to follow up with a PR that drops
the table altogether. It's possible that between this one and that one going
out, we won't be able to fully manage some existing questions because SQLAlchemy
won't know to delete DataSourceItemReferences, so we won't be able to delete
data source items.

We can manage this fine for the moment because there aren't any users. In a
production use case we'd have to do this differently.