Refactor app-markdown-editor to context-aware-editor
Combine the existing `app-markdown-editor`, which can be used to insert a
markdown toolbar above textareas, with a new highlighting and contextualising
overlay. This is currently used only for the guidance textareas for questions
and groups. It contains a set of buttons for inserting markdown for headings,
links and lists into the textarea.

The new contextualising/highlighting module also attaches to a set of question
fields (text, hint, and guidance) and converts context references to
human-readable equivalents that are highlighted to visually identify them.

The new module has some complex functionality to manage this behaviour. To set
it up, we pass a map of context references to their human labels, and point it
at a text area. When initialising, it duplicates the textarea and hices the
original one. The original textarea contains the raw text, which contains
context references like `((q_1234))`. The visible textarea shows human-readable
labels for those references to help form designers understand what they're
seeing. For example, it might show `((lead contact name))` with a yellow
highlight. The duplicated textarea also has an overlay div that adds
highlighting around these references. This level of complexity is required for
the functionality:

The overlay div gets a copy of the text from the visible textarea and is exactly
overlaid on the textarea; text in the div is invisible, except for the
human-readable labels which get a yellow highlight. This ends up overlaying on
the textarea and looks as if it's a background for the human reference.

We need a duplicate textarea because the text in the text areas is different;
one contains raw references and one contains human readable labels. This means
that the length of text between the two changes, so we can't use the
highlighting overlay on the raw textarea while still showing the human labels
instead of the raw IDs.

The hidden textarea retains the name of the field so that it's submitted with
any form it's in; the visible textarea retains the ID of the field so that
anchors/etc target it (eg from error summaries).