FSPT-781: Integrate add context flow for GreaterThan
As the first step in allowing context data insertion into conditions and
validation, we've implemented this flow for the GreaterThan managed
expression. This makes use of the existing `add context data` flow used
in question text, question hints and guidance and the same session
storage pattern.

This was very much a "can we make this work" pass and there are a number
of things that need fixing up eg:
- hiding the expression string field on the managed expression form/
hiding the integer field on the expression form, depending on the value
being used
- clearing the integer value if context data is added
- not mapping things with endswith("_add_context") in quite so many places
- generally a fair bit of tidying up

Potential refactors:
- split out the ManagedExpression form logic from the pydantic models
- split out inline conditional forms onto separate pages
- pass some of the key IDs as URL params (eg. question_id, expression_id,
depends_on_question_id) instead of in the session context

Next steps:
- Add similar functionality for the other relevant ManagedExpression
types (LessThan, Between, and all Date ManagedExpression types)
- Update _validate_and_sync_expression_references to actually
create a new row in the ComponentReference table linking the question
that has the condition/validation and the question it's pulling the
value from
- Add context aware highlighting to the expression reference field and
generally interpolate anywhere that might show the safe_qid

Passing the question safe_qid to the expression engine just works though
so that's a bonus.