FSPT-883: update get answer for question to read at index
If a consumer is reading the answer to an add another question it must
provide which entry it wants to read - for now that puts the
expectation/ requirement for iterating over all entries in an add
another list on the reader/ consumer.

Adds sensible checks to make sure that an index is always provided if
its in an add another context.

Unlike the answer to a question being missing if a consumer has asked
for an entry at an index that doesn't exist this will raise an exception
as the consumer should only be iterating over known entries (vs.
checking if something has been answered which could be None).

This commit needed to update consumers of the "get_answer_for_question"
method that are called whenever a submission helper is instantiated -
otherwise a submission helper will fail under test. These consumers
aren't yet needed for the functionality of this change.

A following commit will update the status helper methods:
- checking visible questions for a form/ group (specifically checks at index if
  is add another and its provided)
- all questions are answered for form (checks at each index for each
  entry)

A following commit will then update the consumers for:
- CSV export
- JSON export
- submission viewer

And finally we'll hook in actually configuring and updating add another
list data in:
- form runner
- form builder