FSPT-781: Only show relevant expression fields
When we're setting expressions up now, you can either have a static value or use
some dynamic data from elsewhere in the system (eg another question). Our form
has separate fields for the static data and the expression statement. We only
want to show one or the other. The expression field currently can only be
populated by using the journey; when returning from that journey it will have
data. So if that field has data, we'll show it and hide the static value form
field. And vice versa.